Reposition quick DM launcher above scroll controls
